+++ radial_menu_model/CHANGELOG.rst
@@ -5,6 +5,7 @@ Changelog for package radial_menu_model
 0.4.0 (2020-XX-XX)
 ------------------
 * Support attributes 'rows' and 'row' to allow null items
+* Use std::shared_ptr<> instead of boost::shared_ptr<>
 
 0.3.4 (2020-08-14)
 ------------------